Move models to TwoPercent namespace
