+++ packages/two_percent/app/models/two_percent/application_record.rb
@@ -1,0 +1,8 @@
+# frozen_string_literal: true
+
+module TwoPercent
+  # @private
+  class ApplicationRecord < ActiveRecord::Base
+    self.abstract_class = true
+  end
+end
